Initial commit.
So far it can get the list of start menu tiles and resolve some lnk shortcuts,
but not all of them as it needs to expand some GUID paths and then read in
the information from the exe files and any manifest files that may be present.

That and the UI doesn't really exist yet.